Reorganize wasm-client-components modules
