+++ crates/cosmos/cosmos-relayer/src/chain/checks.rs
@@ -1,8 +1,9 @@
+use hermes_cli_components::any_client::contexts::any_counterparty::AnyCounterparty;
 use hermes_relayer_components::chain::traits::queries::client_state::CanQueryClientState;
 use hermes_test_components::chain::traits::messages::ibc_transfer::CanBuildIbcTokenTransferMessage;
 use hermes_test_components::chain::traits::queries::balance::CanQueryBalance;
 use hermes_test_components::chain::traits::transfer::ibc_transfer::CanIbcTransferToken;
-use hermes_wasm_client_components::wasm::contexts::wasm_counterparty::WasmCounterparty;
+use hermes_wasm_client_components::contexts::wasm_counterparty::WasmCounterparty;
 
 use crate::contexts::chain::CosmosChain;
 
@@ -11,6 +12,7 @@ pub trait CheckCosmosChainImpls:
     + CanIbcTransferToken<CosmosChain>
     + CanBuildIbcTokenTransferMessage<CosmosChain>
     + CanQueryClientState<WasmCounterparty>
+    + CanQueryClientState<AnyCounterparty>
 {
 }
