+++ crates/sovereign/sovereign-client-components/src/cosmos/impls/client/create_client_message.rs
@@ -5,8 +5,8 @@ use hermes_cosmos_relayer::types::error::{BaseError, Error};
 use hermes_relayer_components::chain::traits::message_builders::create_client::CreateClientMessageBuilder;
 use hermes_relayer_components::chain::traits::types::create_client::HasCreateClientPayloadType;
 use hermes_relayer_components::chain::traits::types::message::HasMessageType;
-use hermes_wasm_client_components::wasm::types::messages::client::consensus::WasmConsensusState;
-use hermes_wasm_client_components::wasm::types::messages::client::state::WasmClientState;
+use hermes_wasm_client_components::types::client_state::WasmClientState;
+use hermes_wasm_client_components::types::consensus_state::WasmConsensusState;
 use ibc_core::primitives::ToProto;
 use ibc_proto::google::protobuf::Any;
 use prost::Message;
