+++ crates/wasm/wasm-client-components/src/impls/mod.rs
@@ -1,3 +1,2 @@
 pub mod decoders;
-pub mod queries;
 pub mod types;
