+++ crates/wasm/wasm-client-components/src/lib.rs
@@ -1,1 +1,4 @@
-pub mod wasm;
+pub mod contexts;
+pub mod impls;
+pub mod types;
+pub mod utils;
