+++ crates/wasm/wasm-client-components/src/types/consensus_state.rs
@@ -1,7 +1,7 @@
 use ibc_proto::google::protobuf::Any;
 use prost::EncodeError;
 
-use crate::wasm::types::messages::utils::encode::encode_to_any;
+use crate::utils::encode::encode_to_any;
 
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
