+++ crates/wasm/wasm-client-components/src/utils/encode.rs
