--- crates/wasm/wasm-client-components/src/wasm/impls/types/mod.rs
@@ -1,1 +1,0 @@
-pub mod client_state;
