--- crates/wasm/wasm-client-components/src/wasm/mod.rs
@@ -1,3 +1,0 @@
-pub mod contexts;
-pub mod impls;
-pub mod types;
