--- crates/wasm/wasm-client-components/src/wasm/types/messages/client/mod.rs
@@ -1,2 +1,0 @@
-pub mod consensus;
-pub mod state;
