--- crates/wasm/wasm-client-components/src/wasm/types/messages/mod.rs
@@ -1,2 +1,0 @@
-pub mod client;
-pub mod utils;
