--- crates/wasm/wasm-client-components/src/wasm/types/mod.rs
@@ -1,1 +1,0 @@
-pub mod messages;
